Update Craters_1.py
For personal use (mostly) 

Take Robbins 2018 Crater database and filter by Area of Interest and eliminate every crater that's 10km in diameter or less. This is because the resolution of our data is about 4.5 Kilometers on the equator. 

Then create a mask of every remaining crater in the region. (Values are 1=Crater, 0=NonCrater)

Use that to generate the labels as each indidivual image is clipped from Gravity image