Image/Label cropper
It assumes you are giving it a georeferenced image/label pair. It may work if they're not geotiffs, but I'm not sure. 
This will create the number of croppings you wish and augmnent each with the described image augmentation methods. It makes sure that each cropping is within bounds. 
I used this to create my datasets for spherical harmonics gravity disturbance data so if you want to use it you need to change a few things in the naming conventions. 
This superseded my crater extractor